chore: firebase configuration
